Create base test case
