+++ custom/icds/tests/test_custom_recipients.py
@@ -1,29 +1,20 @@
-import uuid
-from casexml.apps.case.mock import CaseBlock
 from corehq.apps.data_interfaces.tests.util import create_case
-from corehq.apps.domain.shortcuts import create_domain
-from corehq.apps.hqcase.utils import submit_case_blocks
 from corehq.apps.locations.tests.util import make_loc, setup_location_types
-from corehq.form_processor.backends.sql.dbaccessors import CaseAccessorSQL
-from corehq.form_processor.tests.utils import use_sql_backend
 from corehq.messaging.scheduling.scheduling_partitioned.models import (
     CaseAlertScheduleInstance,
     CaseTimedScheduleInstance,
 )
 from custom.icds.const import AWC_LOCATION_TYPE_CODE, SUPERVISOR_LOCATION_TYPE_CODE
 from custom.icds.messaging.custom_recipients import mother_person_case_from_ccs_record_case
-from django.test import TestCase
-from xml.etree import ElementTree
+from custom.icds.tests.base import BaseICDSTest
 
 
-@use_sql_backend
-class CustomRecipientTest(TestCase):
+class CustomRecipientTest(BaseICDSTest):
     domain = 'icds-custom-recipient-test'
 
     @classmethod
     def setUpClass(cls):
         super(CustomRecipientTest, cls).setUpClass()
-        cls.domain_obj = create_domain(cls.domain)
         cls.mother_person_case = cls.create_case('person')
         cls.child_person_case = cls.create_case(
             'person',
@@ -54,37 +45,6 @@ class CustomRecipientTest(TestCase):
         cls.ls1 = make_loc('ls1', domain=cls.domain, type=SUPERVISOR_LOCATION_TYPE_CODE)
         cls.awc1 = make_loc('awc1', domain=cls.domain, type=AWC_LOCATION_TYPE_CODE, parent=cls.ls1)
         cls.awc2 = make_loc('awc2', domain=cls.domain, type=AWC_LOCATION_TYPE_CODE, parent=None)
-
-    @classmethod
-    def tearDownClass(cls):
-        CaseAccessorSQL.hard_delete_cases(
-            cls.domain,
-            [
-                cls.mother_person_case.case_id,
-                cls.child_person_case.case_id,
-                cls.child_health_extension_case.case_id,
-                cls.lone_child_health_extension_case.case_id,
-                cls.ccs_record_case.case_id,
-            ]
-        )
-        cls.domain_obj.delete()
-        super(CustomRecipientTest, cls).tearDownClass()
-
-    @classmethod
-    def create_case(cls, case_type, parent_case_id=None, parent_case_type=None, parent_identifier=None,
-            parent_relationship=None):
-
-        kwargs = {}
-        if parent_case_id:
-            kwargs['index'] = {parent_identifier: (parent_case_type, parent_case_id, parent_relationship)}
-
-        caseblock = CaseBlock(
-            uuid.uuid4().hex,
-            case_type=case_type,
-            create=True,
-            **kwargs
-        )
-        return submit_case_blocks(ElementTree.tostring(caseblock.as_xml()), cls.domain)[1][0]
 
     def test_mother_person_case_from_child_health_case(self):
         for cls in (CaseAlertScheduleInstance, CaseTimedScheduleInstance):
